added SaveSnippet Modal
