+++ client/components/modals/modalStyle.js
@@ -1,0 +1,20 @@
+
+
+
+
+
+const modalStyle = () => ({
+    content: {
+    top: "50%",
+    left: "50%",
+    right: "auto",
+    bottom: "auto",
+    marginRight: "-50%",
+    transform: "translate(-50%, -50%)",
+    zIndex: 10,
+    width: "400px",
+    height: "300px",
+    },
+})
+
+export default modalStyle
